Numbers in the DataFrame consistent

+++ cue_tools.py
@@ -208,7 +208,7 @@ def CUEDef(model):
         
             CUE22 = SecretionSummary.loc[L,'flux'].values
         
-    FinalCUE2 = (CUE1 + CUE22)/CUE1
+    FinalCUE2 = float((CUE1 + CUE22)/CUE1)
     
     # Third CUE Definition
     
@@ -219,7 +219,7 @@ def CUEDef(model):
     
     # Fourth CUE Definition
     
-    FinalCUE4 = (BiomassProd)/(BiomassProd - CUE22)
+    FinalCUE4 = float((BiomassProd)/(BiomassProd - CUE22))
     
     FinalCUETab = pd.DataFrame(np.array([[FinalCUE1],[FinalCUE2],[FinalCUE3],[FinalCUE4]]), columns = ['CUE'],index = ['Def 1','Def 2','Def 3','Def 4'])
